Move escape_uri method to a more suitable place
