+++ lib/url_helper.rb
@@ -21,4 +21,14 @@ class UrlHelper
     url.sub(/^http:/i, "")
   end
 
+  DOUBLE_ESCAPED_REGEXP ||= /%25([0-9a-f]{2})/i
+
+  # Prevents double URL encode
+  # https://stackoverflow.com/a/37599235
+  def self.escape_uri(uri, pattern = URI::UNSAFE)
+    encoded = URI.encode(uri, pattern)
+    encoded.gsub!(DOUBLE_ESCAPED_REGEXP, '%\1')
+    encoded
+  end
+
 end
