+++ spec/models/topic_embed_spec.rb
@@ -233,25 +233,4 @@ describe TopicEmbed do
     end
   end
 
-  context ".escape_uri" do
-    it "doesn't escape simple URL" do
-      url = TopicEmbed.escape_uri('http://example.com/foo/bar')
-      expect(url).to eq('http://example.com/foo/bar')
-    end
-
-    it "escapes unsafe chars" do
-      url = TopicEmbed.escape_uri("http://example.com/?a=\11\15")
-      expect(url).to eq('http://example.com/?a=%09%0D')
-    end
-
-    it "escapes non-ascii chars" do
-      url = TopicEmbed.escape_uri('http://example.com/ماهی')
-      expect(url).to eq('http://example.com/%D9%85%D8%A7%D9%87%DB%8C')
-    end
-
-    it "doesn't escape already escaped chars" do
-      url = TopicEmbed.escape_uri('http://example.com/foo%20bar/foo bar/')
-      expect(url).to eq('http://example.com/foo%20bar/foo%20bar/')
-    end
-  end
 end
